The code had a bug where creating the interrupt controller didn't work properly because the number of interrupts to handle in the buffer was passed in as an 8-bit value, and this system supports up to 256 interrupts, so the interrupt controller was created with a buffer size of 0. I added a test to make sure that queues must be at least 1 entry long.
The queue code has been changed internally to use size_t's instead of
uint8_t's since we don't want to limit the queue module to only sizes of
256 or less.

+++ include/queue.h
@@ -6,6 +6,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include <stddef.h>
 
 typedef struct queue_t queue_t;
 
@@ -19,7 +20,7 @@ struct queue_return_data_t
 
 typedef struct queue_return_data_t queue_return_data_t;
 
-queue_t* queue_create(uint8_t queue_size);
+queue_t* queue_create(size_t queue_size);
 void queue_destroy(queue_t* queue);
 
 bool queue_is_empty(queue_t* queue);
